chore(*): clean up with files, remove unused spec files
