+++ src/auth/auth.module.ts
@@ -3,8 +3,9 @@ import {AuthService} from './auth.service';
 import {PassportModule} from '@nestjs/passport';
 import {LocalStrategy} from './local-strategy';
 import {JwtModule} from '@nestjs/jwt';
-import {JwtStrategy} from './jwt-strategy';
+import {JwtStrategy} from './jwt/jwt-strategy';
 import {UsersModule} from '../users/users.module';
+import {AuthController} from './auth.controller';
 
 @Module({
     exports: [AuthService],
@@ -12,6 +13,7 @@ import {UsersModule} from '../users/users.module';
         secret: process.env.JWT_SECRET,
         signOptions: {expiresIn: '60s'}
     })],
-    providers: [AuthService, LocalStrategy, JwtStrategy]
+    providers: [AuthService, LocalStrategy, JwtStrategy],
+    controllers: [AuthController]
 })
 export class AuthModule {}
