+++ src/users/users.module.ts
@@ -1,8 +1,10 @@
 import {Module} from '@nestjs/common';
 import {UsersService} from './users.service';
+import {UsersController} from './users.controller';
 
 @Module({
   providers: [UsersService],
-  exports: [UsersService]
+  exports: [UsersService],
+  controllers: [UsersController]
 })
 export class UsersModule {}
